Fixed Sanitising And Added Paragraph Attribute To Markdown Parsing
Since the Markdown parsing library wraps every string into an HTML paragraph a specific new paragraph class "paragraph-small" had to be introduced to ensure a coherent layout. This attribute is added to all HTML paragraphs in the parsed Markdown data.
The sanitising had to be adapted slightly to allow classes on HTML paragraph elements, and only there.