[Banners] Fix admin banner upload
The previous code allowed only conference owners but not
admins to access the conference a banner should be uploaded to.
The new code fetches the conference regardless of access rights,
since the banner upload itself checks whether the account has
the appropriate rights to write to a conference which boils
down to both admins and conference admins.