[FigureService] Docker mobile upload fix
In a docker environment '/target/universal/stage' is used when resolving '.'
in a relative path. 'java.io.File' seems to properly resolve the relative paths
even in a docker environment, but the third party library scrimage cannot work
with these file descriptors and requires file descriptors created with absolute paths.
  As a workaround the file not found exception is caught and the docker container
root path is used to create file descriptior with absolut paths appropriate for
the docker environment.